Schedule ride component
Added a component that can be used to schedule a ride. It contains a map
on which you can search routes, mark routes with a click and choose date
and time. It will be used by unregistered user home page and passenger
home page. However in the unregistered user home page a lot of
functionality will be restricted. Such as scheduling the ride and
choosing date and time.