Routing for chat component
Addded routing for chat component. Route for chat component for admin is
'admin-chat'. Created method appendMessage in message bubbles component
which adds the passed message to this.messages attribute and scrolls
bubbles container to bottom. Because of that, when chat component is
loaded it automatically scrolls to bottom to the latest message.